Fix a bug and move WOWS code out of TheCreatEntity
