+++ src/components/entity-panel/games/Wows.vue
@@ -1,6 +1,14 @@
 <template>
   <div v-if="clickedItem === 'Add'">
-    <TheCreateEntity :clickedItem="clickedItem" :teams="teams" :shipsData="shipsData" :fields="fields" game="wows" />
+    <TheCreateEntity
+      game="wows"
+      :entities.sync="entities"
+      :clickedItem="clickedItem"
+      :teams="teams"
+      :entityData="shipsData"
+      :fields="fields"
+      :autoCompleteOnChangeHandler="autoCompleteOnChangeHandler"
+    />
   </div>
   <div
     v-else
@@ -10,6 +18,7 @@
 </template>
 
 <script lang="ts">
+import axios from 'axios'
 import { Component, Prop, Vue } from 'vue-property-decorator'
 import { RoomGetters, RoomState, Ship } from '@/store/modules/room'
 import { Getter } from 'vuex-class'
@@ -28,6 +37,8 @@ export default class Wows extends Vue {
   @Prop() private clickedItem!: string;
   @Prop() private teams!: MenuItem[];
   @Getter(`room/${RoomGetters.ROOM_STATE}`) roomState!: RoomState;
+
+  entities: Item[] = []
 
   fields: Field[] = [{
     id: 'shipMainBatteryRange',
@@ -87,28 +98,28 @@ export default class Wows extends Vue {
           text: 'Aircraft Carrier (CV)',
           shortText: 'CV',
           value: 1,
-          image: 'https://glossary-wows-global.gcdn.co/icons//vehicle/types/AirCarrier/standard_84f55678325d4b492215390a7f0b43008f3947ab201502cd979dcf4c37633cf3.png',
+          image: 'https://glossary-wows-global.gcdn.co/icons/vehicle/types/AirCarrier/standard_84f55678325d4b492215390a7f0b43008f3947ab201502cd979dcf4c37633cf3.png',
           tier: 0,
           data: {}
         }, {
           text: 'Battleship (BB)',
           shortText: 'BB',
           value: 2,
-          image: 'https://glossary-wows-global.gcdn.co/icons//vehicle/types/Battleship/standard_01624cacb82f39f77a4e677a7b9fdf4df20dafd61f971f4b2d3e54c3065e2892.png',
+          image: 'https://glossary-wows-global.gcdn.co/icons/vehicle/types/Battleship/standard_01624cacb82f39f77a4e677a7b9fdf4df20dafd61f971f4b2d3e54c3065e2892.png',
           tier: 0,
           data: {}
         }, {
           text: 'Cruiser (CA)',
           shortText: 'CA',
           value: 3,
-          image: 'https://glossary-wows-global.gcdn.co/icons//vehicle/types/Cruiser/standard_874a3bdc3134b8da4fd6f52186f1b2b682f13ef78688732d3016785c0649a424.png',
+          image: 'https://glossary-wows-global.gcdn.co/icons/vehicle/types/Cruiser/standard_874a3bdc3134b8da4fd6f52186f1b2b682f13ef78688732d3016785c0649a424.png',
           tier: 0,
           data: {}
         }, {
           text: 'Destroyer (DD)',
           shortText: 'DD',
           value: 4,
-          image: 'https://glossary-wows-global.gcdn.co/icons//vehicle/types/Destroyer/standard_357acc9fc0e2f7d98f047c99edffad359a8c45f2093024400fef2b9abbaf3a59.png',
+          image: 'https://glossary-wows-global.gcdn.co/icons/vehicle/types/Destroyer/standard_357acc9fc0e2f7d98f047c99edffad359a8c45f2093024400fef2b9abbaf3a59.png',
           tier: 0,
           data: {}
         }, ...this.roomState.game?.ships.map((ship: Ship) => ({
@@ -122,6 +133,18 @@ export default class Wows extends Vue {
       }
     }
     return []
+  }
+
+  async autoCompleteOnChangeHandler (shipItem: Item) {
+    if (shipItem) {
+      this.entities.unshift(shipItem)
+      const response = await axios.get(`https://api.worldofwarships.eu/wows/encyclopedia/shipprofile/?ship_id=${shipItem.value}&application_id=d84a218b4fec37003e799f13777bf880`)
+      const shipData = response.data.data[shipItem.value]
+      const foundNewEntity = this.entities.find((entity: Item) => entity.value === shipItem.value)
+      if (foundNewEntity) {
+        foundNewEntity.data = shipData
+      }
+    }
   }
 }
 </script>
